v2functions/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
requests==2.20.1
feedparser==5.2.1